add a suite of benchmarks
In order to effectively understand the impact of a change to this
codebase to the performance of the product, a suite of benchmarks is
necessary. Two major value recording methods exist - for numeric values
with a fixed value and scale, and for floating-point values. The
benchmarks in this patch are written to focus on a couple key points:

 - Users of this histogram may record values from very disparate
   distributions, some with small differences in scale (say, request
   response times between milliseconds and tens of seconds) and some may
   record values with a vast distributions. The benchmarks should tease
   apart difference in performance that occur due to these usage
   patterns.
 - Generating and storing the values which will ultimately be fed to the
   histogram for recording may take a significant amount of time, so the
   benchmarks should be careful to separate this time from the time
   taken to record the values into buckets.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>