allow creating histograms without lookup tables
In many cases when the number of histogram buckets is low, the sparse
lookup tables take up the vast majority of the memory footprint of the
histogram. This patch adds an option to allow users to create histograms
without lookup tables for buckets. The effect is to drastically reduce
the memory footprint of a histogram while only increasing insert time by
about 20%.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>