introduce a deserialization target without lookups
When users need to explicitly deserialize histogram data while not
allocating memory for lookup tables, `DeserializeWithOptions` can be
used. When the deserialization is implicit to Go's JSON library,
however, a custom type is required in order to ensure that no excess
memory is allocated by reflect (during the JSON deserialization) or
by this library (during histogram data deserialization).

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>